# backend/middleware/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Cloud Engine API",
    "version": "0.0.1",
    "description": "These API endpoint documentations are autogenerated from the backend code."
  },
  "servers": [
    {
      "url": "http://localhost:3000/"
    }
  ],
...
  "paths": {
    "/api/sdk/iam/getAllUsers": {
      "get": {
        "tags": [
          "SDK - IAM"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/sdk/iam/getUserByEmail": {
      "post": {
        "tags": [
          "SDK - IAM"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/sdk/iam/createUser": {
      "post": {
        "tags": [
          "SDK - IAM"
        ],
        "description": "",
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "example": "any"
                  },
                  "username": {
                    "example": "any"
                  },
                  "rdsLink": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/sdk/iam/delete": {
      "delete": {
        "tags": [
          "SDK - IAM"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/sdk/iam/replace": {
      "patch": {
        "tags": [
          "SDK - IAM"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "in": "header",
        "bearerFormat": "JWT"
      }
    }
  }
}